+++ CHANGELOG.md
@@ -7,6 +7,12 @@
 ### **Copyrigth**: 
 *Tiledesk SRL*
 
+# 1.39.0-rc8
+- **bug-fix**: fixed bugs in action ai-condition
+
+# 1.39.0-rc7
+- **added**: disabled property for mat-slider in ai-settings section
+
 # 1.39.0-rc6
 - **added**: added action ai-condition